requirements.txt
mee6_py_api==0.0.1
tatsu.py==1.0.4
discord.py==2.0.0
sqlitedict==1.7.0
numpy==1.21.4
Random_Word==1.0.7
pandas==1.3.4
discord==1.7.3
GitPython==3.1.27
python-dotenv==0.20.0
tatsu==5.8.3
discord-pretty-help==1.3.4